ver -- 1.0.4

+++ src/main/java/KT/PLIP/population/domain/LongForeigner.java
@@ -15,7 +15,7 @@ public class LongForeigner {
     private Long id;
     
     @Column(name = "tot_lvpop_co")
-    private Integer totLvpopCo;
+    private Double totLvpopCo;
     
     @Column(name = "china_staypop_co")
     private Double chinaStaypopCo;
