+++ src/main/java/KT/PLIP/population/domain/TempForeigner.java
@@ -15,7 +15,7 @@ public class TempForeigner {
     private Long id;
     
     @Column(name = "tot_lvpop_co")
-    private Integer totLvpopCo;
+    private Double totLvpopCo;
     
     @Column(name = "china_staypop_co")
     private Double chinaStaypopCo;
